data-engineering/zoomcamp/final-project/pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.4",
    "faker>=40.4.0",
    "google-api-core>=2.29.0",
    "google-cloud-bigquery>=3.40.0",
    "pandas>=3.0.0",
    "polars>=1.38.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.46",
]
